pylint and other cleanup
